Refactor access control and db schema

+++ backend/app/controllers/practitioner/exercises.js
@@ -1,34 +1,61 @@
 const prisma = require('../prisma-client');
 
 const create = async (ctx) => {
+
+  // Check if req planId belongs to practitioner
+  if (ctx.practitioner.plans.every((plan) => {
+    plan.id != ctx.request.body.planId
+  })) {
+    ctx.status = 401;
+    return
+  }
+
   const exercise = await prisma.exercise.create({
     data: {
       ...ctx.request.body
     }
   });
 
-  ctx.body = {
-    data: exercise
-  };
+  ctx.body = exercise;
   ctx.status = 200
 }
 
 const get = async (ctx) => {
   const id = parseInt(ctx.params.id);
+
   const exercise = await prisma.exercise.findUnique({
     where: {
       id: id
     }
   });
 
-  ctx.body = {
-    data: exercise
-  };
+  if (ctx.practitioner.plans.every((plan) => {
+    plan.id != exercise.planId
+  })) {
+    ctx.status = 401;
+    return
+  }
+
+  ctx.body = exercise;
   ctx.status = 200;
 }
 
 const update = async (ctx) => {
   const id = parseInt(ctx.params.id);
+
+  const exercise = await prisma.exercise.findUnique({
+    where: {
+      id: id
+    }
+  });
+
+  if (ctx.practitioner.plans.every((plan) => {
+    plan.id != exercise.planId
+  })) {
+    ctx.status = 401;
+    return
+  }
+
   const updateExercise = await prisma.exercise.update({
     where: {
       id: id
@@ -36,15 +63,27 @@ const update = async (ctx) => {
     data: ctx.request.body
   });
 
-  ctx.body = {
-    data: updateExercise
-  }
+  ctx.body = updateExercise;
   ctx.status = 200;
 }
 
 const destroy = async (ctx) => {
   const id = parseInt(ctx.params.id);
-  const exercise = await prisma.exercise.delete({
+
+  const exercise = await prisma.exercise.findUnique({
+    where: {
+      id: id
+    }
+  });
+
+  if (ctx.practitioner.plans.every((plan) => {
+    plan.id != exercise.planId
+  })) {
+    ctx.status = 401;
+    return
+  }
+
+  await prisma.exercise.delete({
     where: {
       id: id
     }
